services/embeddings/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
